Remove extraneous helper methods and move helpers to base of tests
Move tests.app.helpers > tests.helpers
Remove assert_response_status in favour of normal assert statement
Consolidate common setup method of TestSearchEndpoint and TestSearchResultsOrdering
Create BaseSearchTestWithServices
Remove create_services now only used in the consolidated setup
Move services_mapping and make_services to tests.helpers
